Fixed a bug where two sites are treated as the same if they are in the same node and the same type. This would lead to issues when more than one site of ine type was on one node. Now ids are compared.
Task sites will now display the task they are representing in their fields

Removed ability for actors to deposit resources that are on the ground in the same node as the site, resources must now be in the actors inventory and the actor must be in the node of the site.

Actors will not go idle if they reach exactly the minimum effort to complete a site, causing them to never be able to finish a site.